Factor out duplicated clientPath type.
This was previously defined privately in both server and the main capnp
package.
